pyproject.toml
[tool.poetry]
name = "course-work-healthy-habits"
version = "0.1.0"
description = ""
authors = ["Viktor Nikulaev <viktor_1611@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "^24.10.0"
psycopg2-binary = "^2.9.10"
ipython = "^8.29.0"
flake8 = "^7.1.1"
python-dotenv = "^1.0.1"
django = "^5.1.3"
djangorestframework = "^3.15.2"
rest-framework-simplejwt = "^0.0.2"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"
django-filter = "^24.3"
celery = "^5.4.0"
eventlet = "^0.38.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
